+++ RELEASE-NOTES.md
@@ -5,6 +5,7 @@
 * #15 Make RmqConnectionManager injectable
 * #16 Injectable RmqTopologyManager
 * #17 Split RmqChannelFactory functions between RmqChannelManager and RmqTopologyManager
+* #20 Queue config templates for sub endpoints (short-lived queues)
 
 ## 2.0.RC1